requirements.txt
Flask==3.0.2
numpy==1.26.4
Pillow==10.3.0
Requests==2.31.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
